+++ src/Simplify.DI.Integration.Microsoft.Extensions.DependencyInjection.Tester/Program.cs
@@ -5,10 +5,7 @@ namespace Simplify.DI.Integration.Microsoft.Extensions.DependencyInjection.Teste
 {
 	public class Program
 	{
-		public static void Main(string[] args)
-		{
-			CreateWebHostBuilder(args).Build().Run();
-		}
+		public static void Main(string[] args) => CreateWebHostBuilder(args).Build().Run();
 
 		public static IWebHostBuilder CreateWebHostBuilder(string[] args) =>
 			WebHost.CreateDefaultBuilder(args)