+++ src/Simplify.DI.Integration.Microsoft.Extensions.DependencyInjection.Tester/Setup/ServiceCollectionIocRegistrations.cs
@@ -4,9 +4,7 @@ namespace Simplify.DI.Integration.Microsoft.Extensions.DependencyInjection.Teste
 {
 	public static class ServiceCollectionIocRegistrations
 	{
-		public static void Register(this IServiceCollection services)
-		{
+		public static IServiceCollection RegisterAll(this IServiceCollection services) =>
 			services.AddScoped<Dependency2>();
-		}
 	}
 }