+++ src/Simplify.DI.Provider.Autofac.Tests/AutofacDIContainerProviderTests.cs
@@ -16,6 +16,25 @@ namespace Simplify.DI.Provider.Autofac.Tests
 			_provider = new AutofacDIProvider();
 		}
 
+		[Test]
+		public void BeginLifetimeScope_WithOnBeginLifetimeScopeEventSubscribed_EventFired()
+		{
+			// Arrange
+
+			var eventFired = false;
+
+			_provider.OnBeginLifetimeScope += delegate
+			{
+				eventFired = true;
+			};
+
+			// Act
+			using var scope = _provider.BeginLifetimeScope();
+
+			// Assert
+			Assert.IsTrue(eventFired);
+		}
+
 		#region Existence tests
 
 		[Test]