+++ src/Simplify.DI/IDIContextHandler.cs
@@ -6,6 +6,11 @@
 	public interface IDIContextHandler
 	{
 		/// <summary>
+		/// Occurs when the lifetime scope is opened
+		/// </summary>
+		event BeginLifetimeScopeEventHandler OnBeginLifetimeScope;
+
+		/// <summary>
 		/// Begins the lifetime scope.
 		/// </summary>
 		/// <returns></returns>