+++ src/Simplify.DI.Provider.Autofac/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [1.1.1] - 2021-10-16
+
+### Dependencies
+
+- Autofac bump to 6.3.0 (PR#311)
+- Simplify.DI bump to 4.1.1
+
+
+## [1.1.0] - 2021-07-18
+
+### Dependencies
+
+- Simplify.DI bump to 4.1
+
 ## [1.0.1] - 2021-04-22
 
 ### Dependencies