+++ src/Simplify.DI/CHANGELOG.md
@@ -1,5 +1,33 @@
 # Changelog
 
+## [4.1.3] - 2021-11-03
+
+### Changed
+
+- DryIOC container verification when Singleton delegate depends on Scoped will throw exceptions
+
+### Dependencies
+
+- Internal DryIoc bump to 4.8.4 (PR#316)
+
+## [4.1.2] - 2021-10-25
+
+### Dependencies
+
+- Internal DryIoc bump to 4.8.3 (PR#314)
+
+## [4.1.1] - 2021-10-15
+
+### Dependencies
+
+- Internal DryIoc bump to 4.8.2 (PR#312)
+
+## [4.1.0] - 2021-07-15
+
+### Added
+
+- Event to subscribe on lifetime scope creation (#303)
+
 ## [4.0.20] - 2021-07-06
 
 ### Dependencies